Implemented page frame and page table allocator
The page frame allocator is a bitmap allocator that allocates
the unmapped space between the two ALLOCATOR macros. There is
no free capability yet, but doing this would simply require
walking through the page table to see the physical address and
then updating the bitmap.

The page table allocator is even more simple. Since I have
already reserved space for the page table pages. I just needed
to use a watermark to grab those pages.

This change is being made because the OS needs some sort of
on demand allocation. We are now able to ask for a page frame
to be mapped anywhere in virtual address space without worrying
about physical address layout.