Implemented basic task switching between 2 tasks
There are a few things that I still want to add
to basic task switching. Right now, I clobber
on all of the registers (except eip and ebp) and
I am not sure how viable this is. The next step
will be to move away from this horribly broken
form of cooperative multitasking. Scheduling land
here we come!