Added basic file support.
VFS layer relies on a basic function to pointer mechanism.
Implemented the initrd mechanism that reads files from a premade
custom ramdisk file format. I will probably need to make changes
to this format in the future as it is not very robust and has
a not so good header. Implemented ramdisk read and write. I will
probably implement the ramdisk filesystem first because it is the
simplest. There are still a few things I want to do for the
ramdisk filesystem:

- Add support for directories in the initrd. Right now everything
  goes into root.
- Add support for ramdisk_open() and ramdisk_close()
- Add an interator or some interface for fetching files from a
  directory.